pull consentstate types out
