+++ src/ccpa/types/CCPAConsentState.ts
@@ -1,0 +1,3 @@
+export interface CCPAConsentState {
+	doNotSell: boolean;
+}
